requirements.txt
python-telegram-bot==13.15
python-dotenv==0.20.0
requests==2.27.1
